Changed project layout to OOP
